test: split recommendations scenarios
