+++ frontend/tests/recommendations/PLAN.md
@@ -1,0 +1,13 @@
+# 推奨機能テスト分割計画
+
+- 初期ロードとフォールバック
+  - 保存済み地域の初回フェッチ
+  - 最新API成功時のレガシーAPI抑止
+  - fetchRecommendations 失敗時のレガシーフォールバック
+  - 初期ロードのフォールバック検証
+  - 週入力正規化の確認
+- 地域変更と競合制御
+  - 地域選択と週入力による手動フェッチ
+  - 地域変更による即時フェッチとテーブル更新
+- お気に入り並び替え
+  - お気に入り優先描画とトグルの永続化
