+++ frontend/tests/recommendations/initialLoad.test.tsx
@@ -1,0 +1,129 @@
+import '@testing-library/jest-dom/vitest'
+import { cleanup, screen, waitFor } from '@testing-library/react'
+import { afterEach, beforeEach, describe, expect, it, type MockInstance } from 'vitest'
+
+import {
+  fetchCrops,
+  fetchRecommend,
+  fetchRecommendations,
+  renderApp,
+  storageState,
+} from '../utils/renderApp'
+import {
+  createItem,
+  createRecommendResponse,
+  defaultCrops,
+  setupRecommendationsTest,
+} from '../utils/recommendations'
+
+describe('App recommendations / 初期ロードとフォールバック', () => {
+  let useRecommendationsSpy: MockInstance
+
+  beforeEach(async () => {
+    ;({ useRecommendationsSpy } = await setupRecommendationsTest())
+  })
+
+  afterEach(() => {
+    useRecommendationsSpy.mockRestore()
+    cleanup()
+  })
+
+  it('保存済み地域で初回フェッチされる', async () => {
+    storageState.region = 'cold'
+    fetchCrops.mockResolvedValue([])
+    fetchRecommendations.mockResolvedValue(
+      createRecommendResponse({ region: 'cold' }),
+    )
+
+    await renderApp()
+
+    await waitFor(() => {
+      expect(fetchRecommendations).toHaveBeenNthCalledWith(1, 'cold', '2024-W30')
+    })
+    expect(useRecommendationsSpy).toHaveBeenCalled()
+  })
+
+  it('最新API成功時はレガシーAPIを呼び出さない', async () => {
+    fetchCrops.mockResolvedValue(defaultCrops.slice(0, 2))
+    fetchRecommendations.mockResolvedValue(createRecommendResponse())
+    fetchRecommend.mockResolvedValue(createRecommendResponse())
+
+    await renderApp()
+
+    await waitFor(() => {
+      expect(fetchRecommendations).toHaveBeenCalledTimes(1)
+      expect(fetchRecommendations).toHaveBeenCalledWith('temperate', '2024-W30')
+    })
+    expect(fetchRecommend).not.toHaveBeenCalled()
+  })
+
+  it('fetchRecommendations が失敗しても fetchRecommend で初期描画される', async () => {
+    fetchCrops.mockResolvedValue(defaultCrops.slice(0, 2))
+    fetchRecommendations.mockRejectedValueOnce(new Error('unexpected error'))
+    fetchRecommend.mockResolvedValue(
+      createRecommendResponse({
+        items: [
+          createItem({ crop: '春菊', source: 'legacy' }),
+        ],
+      }),
+    )
+
+    await renderApp()
+
+    await waitFor(() => {
+      expect(fetchRecommend).toHaveBeenCalledWith({ region: 'temperate', week: '2024-W30' })
+    })
+    expect(screen.getByText('春菊')).toBeInTheDocument()
+  })
+
+  it('初期ロードで fetchRecommendations が失敗したら fetchRecommend にフォールバックする', async () => {
+    fetchCrops.mockResolvedValue(defaultCrops.slice(0, 2))
+    fetchRecommendations.mockRejectedValueOnce(new Error('network error'))
+    fetchRecommend.mockResolvedValue(
+      createRecommendResponse({
+        items: [
+          createItem({ crop: '春菊', source: 'legacy' }),
+        ],
+      }),
+    )
+
+    await renderApp()
+
+    await waitFor(() => {
+      expect(fetchRecommendations).toHaveBeenCalledTimes(1)
+      expect(fetchRecommend).toHaveBeenCalledTimes(1)
+    })
+    expect(fetchRecommendations).toHaveBeenNthCalledWith(1, 'temperate', '2024-W30')
+    expect(fetchRecommend).toHaveBeenCalledWith({ region: 'temperate', week: '2024-W30' })
+    expect(screen.getByText('春菊')).toBeInTheDocument()
+  })
+
+  it('週入力は normalizeIsoWeek で揃えてAPIへ送られる', async () => {
+    fetchCrops.mockResolvedValue(defaultCrops.slice(0, 2))
+    fetchRecommendations.mockImplementation(async (region, week) => {
+      const resolvedWeek = week ?? '2024-W30'
+      return createRecommendResponse({
+        week: resolvedWeek,
+        region,
+        items: [createItem({ crop: '春菊' })],
+      })
+    })
+
+    const { user } = await renderApp()
+
+    await waitFor(() => {
+      expect(fetchRecommendations).toHaveBeenLastCalledWith('temperate', '2024-W30')
+    })
+
+    const select = screen.getByLabelText('地域')
+    const weekInput = screen.getByLabelText('週')
+    await user.selectOptions(select, '寒冷地')
+    await user.clear(weekInput)
+    await user.type(weekInput, '2024w33')
+    await user.click(screen.getByRole('button', { name: 'この条件で見る' }))
+
+    await waitFor(() => {
+      expect(fetchRecommendations).toHaveBeenLastCalledWith('cold', '2024-W33')
+    })
+  })
+})
